feat(pagination): move Pages tests back to one file
