+++ packages/pagination/tests/Pages.test.js
@@ -3,20 +3,60 @@ import { render, cleanup, fireEvent } from 'react-testing-library';
 
 import 'jest-dom/extend-expect';
 
-import Pages from '../PaginationControls/Pages';
+import Pages, { defaultButtonText } from '../PaginationControls/Pages';
 
-import { testPagesRender } from './pagesTestFn';
+const pagesTestId = 'page-selector';
 
 describe('Pages', () => {
   afterEach(cleanup);
   let mockFn;
+  let baseProps;
   beforeEach(() => {
     mockFn = jest.fn();
+    baseProps = {
+      onPageChange: jest.fn(),
+      pageCount: 5,
+    };
   });
 
-  testPagesRender({
-    Component: Pages,
+  test('renders as expected', () => {
+    const { getByTestId } = render(<Pages {...baseProps} />);
+    expect(getByTestId(pagesTestId)).toBeDefined();
   });
+
+  test('onPageChange is required', () => {
+    const useProps = {
+      ...baseProps,
+      onPageChange: undefined,
+    };
+    expect(() => {
+      render(<Pages {...useProps} />);
+    }).toThrow();
+  });
+
+  test('pageCount must be defined', () => {
+    const useProps = {
+      ...baseProps,
+      pageCount: undefined,
+    };
+    expect(() => {
+      render(<Pages {...useProps} />);
+    }).toThrowError(/must define pageCount or totalCount and itemsPerPage/);
+  });
+
+  // TODO: figure out how to get this test working to validate props
+  // test('pageCount, totalCount, and itemsPerPage must be positive if defined', () => {
+  //   ['pageCount', 'totalCount', 'itemsPerPage'].forEach(propName => {
+  //     const errorCheck = new RegExp(`${propName} must be a positive number`);
+  //     [-2, -1].forEach(i => {
+  //       expect(() => {
+  //         const props = { ...baseProps, [propName]: i };
+  //         // console.log(`${propName}: ${props[propName]}`);
+  //         render(<Pages {...props} />);
+  //       }).toThrowError(errorCheck);
+  //     });
+  //   });
+  // });
 
   test('should add correct classes for size,align,unstyled', () => {
     const { container, rerender } = render(
@@ -56,6 +96,60 @@ describe('Pages', () => {
 
         rerender(<Pages onPageChange={mockFn} {...testProps} pageCount={5} />);
         expect(container.firstChild).toHaveClass(expected);
+      });
+    });
+  });
+
+  describe('Navigation buttons', () => {
+    test('renders nav buttons by default', () => {
+      const { getByLabelText } = render(<Pages {...baseProps} />);
+      [
+        { label: 'First', text: defaultButtonText.firstBtn },
+        { label: 'Previous', text: defaultButtonText.prevBtn },
+        { label: 'Next', text: defaultButtonText.nextBtn },
+        { label: 'Last', text: defaultButtonText.lastBtn },
+      ].forEach(testCase => {
+        expect(getByLabelText(testCase.label)).toHaveTextContent(testCase.text);
+      });
+    });
+
+    test('renders nav buttons with expected text', () => {
+      const { queryByLabelText, rerender } = render(<Pages {...baseProps} />);
+
+      [
+        {
+          label: 'First',
+          prop: 'firstBtn',
+          testValues: ['hello world', 'First', 'First Page', false],
+        },
+        {
+          label: 'Previous',
+          prop: 'prevBtn',
+          testValues: ['Previous', 'Previous Page', false],
+        },
+        {
+          label: 'Next',
+          prop: 'nextBtn',
+          testValues: ['Next', 'Next Page', false],
+        },
+        {
+          label: 'Last',
+          prop: 'lastBtn',
+          testValues: ['Last', 'Last Page', false],
+        },
+      ].forEach(testCase => {
+        testCase.testValues.forEach(testText => {
+          rerender(<Pages {...baseProps} {...{ [testCase.prop]: testText }} />);
+
+          const elm = queryByLabelText(testCase.label);
+
+          const test = expect(elm);
+          if (!testText) {
+            test.toBeNull();
+          } else {
+            test.toHaveTextContent(testText);
+          }
+        });
       });
     });
   });
